requirements.txt
cryptography
PyJWT
requests
django
django-rest-framework
